requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.2.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.8
iniconfig==2.0.0
jalali_core==1.0.0
jdatetime==5.0.0
kombu==5.4.0
multidict==6.0.5
packaging==24.1
phonenumbers==8.13.44
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.8
requests==2.32.3
requests-mock==1.12.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.7
